+++ tests/testthat/test-checks.R
@@ -33,3 +33,27 @@ describe("wcvp_check_version()", {
     expect_warning(wcvp_check_version(silent = FALSE), "Could not check")
   })
 })
+
+describe("wcvp_validate_data()", {
+  it("validates correct data successfully", {
+    expect_true(wcvp_validate_data(silent = TRUE))
+  })
+
+  it("identifies invalid names dataset dimensions", {
+    empty_dist <- wcvp_checklist_distribution[0, ]
+    expect_false(wcvp_validate_data(names = data.frame(), distribution = empty_dist, silent = TRUE))
+    expect_warning(
+      expect_warning(
+        wcvp_validate_data(names = data.frame(), distribution = empty_dist, silent = FALSE),
+        "Names dataset dimensions mismatch"
+      ),
+      "Distribution dataset dimensions mismatch"
+    )
+  })
+
+  it("identifies invalid distribution dataset dimensions", {
+    expect_false(wcvp_validate_data(distribution = data.frame(), silent = TRUE))
+    expect_warning(wcvp_validate_data(distribution = data.frame(), silent = FALSE), "dimensions mismatch")
+  })
+})
+